Update and rename cp3_kb406510.R to Pricing European option in 1.5 model of stochastic volatility
The stock price follows the 3/2 model of stochastic volatility. This program gives function which calculates the price of European call in this model by the Monte Carlo method. The function computes S_T by calculating trajectories of S_t by the Euler-Maruyama scheme and trajectories of V_t (volatility process in the 3/2 model) by Milstein scheme.